requirements.txt
# Core dependencies
aiohttp>=3.9.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0

# Exchange SDKs for order signing
lighter-sdk>=1.0.0
x10-python-trading-starknet>=0.1.0
